backend/services/ai-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
chromadb==0.4.22
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.2.0
llama-cpp-python==0.2.37
sse-starlette==2.0.0
python-multipart==0.0.6
numpy==1.26.3
